requirements.txt
asgiref==3.10.0
audioop-lts==0.2.2
audioread==3.1.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
decorator==5.2.1
Django==5.2.7
django-csp==4.0
django-filter==25.2
django-taggit==6.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.2
fsspec==2025.12.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.1.2
mysql-connector-python==9.5.0
mysqlclient==2.2.7
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
packaging==25.0
pillow==12.0.0
platformdirs==4.5.1
pooch==1.8.2
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.3
setuptools==70.2.0
soundfile==0.13.1
soxr==1.0.0
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.1+cpu
torchaudio==2.9.1+cpu
torchvision==0.24.1+cpu
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.4
